requirements.txt
fastapi[all]==0.88.0
uvicorn[standard]==0.20.0
pydantic
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
sqlalchemy_utils
pymysql
python-dotenv